fix(app): Refactorización y añadidos validación en backend
