+++ auth.config.ts
@@ -2,7 +2,7 @@ import type { NextAuthConfig } from "next-auth";
 import Credentials from "next-auth/providers/credentials";
 import { getUserByEmail } from "./data/user";
 import bcryptjs from "bcryptjs";
-import { signInSchema } from "./lib/zod";
+import { signInSchema } from "./lib/schemas/auth/login.schema";
 
 export default {
   providers: [
