+++ lib/schemas/auth/register.schema.ts
