+++ lib/schemas/auth/updateProfile.schema.ts
@@ -1,0 +1,7 @@
+import z from "zod";
+
+export const formSchema = z.object({
+  name: z.string().min(1, "Nombre obligatorio"),
+  firstSurname: z.string().min(1, "Primer apellido obligatorio"),
+  secondSurname: z.string().optional(),
+});
